+++ KnilKit/Model/Errors.swift
@@ -11,4 +11,18 @@ import Foundation
 public enum KnilKitError: Error {
     case invalidURLString(String)
     case noData
+    case cannotFetchFile
+}
+
+extension KnilKitError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case .invalidURLString(let urlString):
+            return "Invalid URL: \(urlString)."
+        case .noData:
+            return "Sorry, no data."
+        case .cannotFetchFile:
+            return "Cannot fetch the file."
+        }
+    }
 }