+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Universal Links plural
 - Link composer use form sheet for iPad #7
+- Improve error messages #5
 
 ## [1.0.0(7)] - 2018-08-06
 ### Fixed